+++ sfja/src/components/Upload/PreviewBlankForm.js
@@ -27,14 +27,11 @@ class PreviewBlankForm extends React.Component {
         form_id: id,
         form_name: name,
       };
-      fetch('/updateFormName', {
+      fetch('http://127.0.0.1:5000/updateFormName', {
         method: 'POST',
         headers: {'Content-Type': 'application/json'},
         body: JSON.stringify(body),
-        // eslint-disable-next-line arrow-parens
-      })
-          .then((res) => res.text())
-          .then((res) => console.log(res));
+      });
     }
 
     // eslint-disable-next-line require-jsdoc
@@ -50,7 +47,7 @@ class PreviewBlankForm extends React.Component {
           </Button>
           <br/>
           <TextField onChange={(e) => {
-            this.setState({value: e.target.newName});
+            this.setState({value: e.target.value});
           }} id="name-field" defaultValue={this.props.parentData.name}>
           </TextField>
           <br />
